updated path from src/css/main.css to dist/css/main.min.css
the starter.html we clone is created with all the paths we are told to create in lines 124 through 127 expect program.html and rand.html.  There is no need to create src/css/main.css because we are already running gulp and should compile from .scss to main.min.css.  If this is not correct please, let me know.